Move usage of classes from optional dependency to their own class
